zbieranie statystyk porównań i zmian węzłów

+++ 4/tree/structs/Stats.java
@@ -1,7 +1,7 @@
 package tree.structs;
 
 public class Stats {
-    public int keyComp = 0;
-    public int nodeComp = 0;
-    public int nodeChange = 0;
+    public long keyComp = 0;
+    public long nodeComp = 0;
+    public long nodeChange = 0;
 }
